# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prompt API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/prompt/{identifier}": {
            "post": {
                "tags": [
                    "Prompts"
                ],
                "summary": "Voer een prompt uit met gegeven parameters",
                "operationId": "b99428bede9f8a66e503406d2d376608",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Identifier van de prompt die uitgevoerd moet worden",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parameters"
                                ],
                                "properties": {
                                    "parameters": {
                                        "description": "Parameters die in zowel de prompt template als system template worden ingevuld",
                                        "type": "object",
                                        "example": {
                                            "tone": "enthousiaste",
                                            "industry": "elektronica",
                                            "product": "Smart TV"
                                        }
                                    },
                                    "temperature": {
                                        "description": "ChatGPT temperature (0.0 - 1.0)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.7
                                    },
                                    "max_tokens": {
                                        "description": "Maximum aantal tokens in de response",
                                        "type": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succesvolle response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "description": "De gegenereerde tekst van ChatGPT",
                                            "type": "string",
                                            "example": "Dit is een voorbeeld response van ChatGPT."
                                        },
                                        "prompt_id": {
                                            "description": "ID van de gebruikte prompt",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "system_message": {
                                            "description": "De gebruikte system message (indien aanwezig)",
                                            "type": "string",
                                            "example": "Je bent een enthousiaste copywriter die gespecialiseerd is in elektronica producten."
                                        },
                                        "usage": {
                                            "description": "Token usage statistieken",
                                            "properties": {
                                                "prompt_tokens": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "completion_tokens": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total_tokens": {
                                                    "type": "integer",
                                                    "example": 30
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ongeldige aanvraag",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Niet alle prompt parameters zijn ingevuld"
                                                },
                                                "prompt": {
                                                    "type": "string",
                                                    "example": "Dit is een {missing_parameter}"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Niet alle system parameters zijn ingevuld"
                                                },
                                                "system_template": {
                                                    "type": "string",
                                                    "example": "Je bent een {missing_parameter} copywriter"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Prompt niet gevonden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Prompt niet gevonden"
                                        },
                                        "identifier": {
                                            "type": "string",
                                            "example": "niet-bestaande-prompt"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "OpenAI service niet beschikbaar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "OpenAI API fout"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service is currently unavailable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Er is een onverwachte fout opgetreden"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Prompts",
            "description": "Prompts"
        }
    ]
}